refactor out expected file result into fixtures
