+++ tests/test_views.py
@@ -1,6 +1,5 @@
 import base64
 import copy
-import xml
 from urllib import parse
 
 import pytest
@@ -26,15 +25,6 @@ from djangosaml2idp.views import (BINDING_HTTP_POST, BINDING_HTTP_REDIRECT,
                                   store_params_in_session)
 
 User = get_user_model()
-
-FILE_PREFIX = "tests/"
-
-
-@pytest.fixture()
-def sample_saml_minimal():
-    with open(FILE_PREFIX + "xml/min/request/sample_saml_request_minimal.xml") as f:
-        expected_result = f.readline()
-    return expected_result
 
 
 @pytest.fixture()
